requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
websockets>=12.0
pytest>=7.0.0
pytest-cov>=4.0.0
